변경 사항: `Gateway` 영역 수정

+++ src/main/java/com/nhnacademy/gateway/gateway_info/controller/GatewayController.java
@@ -28,7 +28,7 @@ public class GatewayController {
     }
 
     @PostMapping
-    public ResponseEntity<Integer> registerGateway(
+    public ResponseEntity<Long> registerGateway(
             @Validated @RequestBody GatewayRegisterRequest request
     ) {
         return ResponseEntity
